create trade_pricing module
